To CPU issue?

+++ code/Seq2SeqModel.py
@@ -38,7 +38,7 @@ class Seq2SeqModel(LightningModule):
     def training_step(self, batch):
 
         x, y_dec_ip, y_dec_op, X_len, _, _ = batch
-        X_len = X_len.cpu()
+        X_len = X_len.cpu().long()
         logits = self(x, X_len, y_dec_ip)  # (batch, tgt_len, vocab_size)
         ## reshaping to match the required shape of (N,C) for logits
         ## and (N,) for label
@@ -63,7 +63,7 @@ class Seq2SeqModel(LightningModule):
     def validation_step(self, batch):
 
         x, y_dec_ip, y_dec_op, X_len, _, _ = batch
-        X_len = X_len.cpu()
+        X_len = X_len.cpu().long()
         logits = self(x, X_len, y_dec_ip)  # (batch, tgt_len, vocab_size)
         ## reshaping to match the required shape of (N,C) for logits
         ## and (N,) for label
